Drop the llvm-external-projects version of the transform interpreter
This code turned out to be dead and has been replaced in IREE core, and
related functionality will be going away in the next integrate.
