+++ llvm-external-projects/iree-dialects/include/iree-dialects/Dialect/LinalgTransform/Passes.h
@@ -12,7 +12,6 @@ namespace mlir {
 namespace linalg {
 namespace transform {
 
-void registerTransformDialectInterpreterPass();
 void registerDropSchedulePass();
 
 } // namespace transform
@@ -22,10 +21,5 @@ void registerDropSchedulePass();
 namespace mlir {
 class Pass;
 
-// Pass to schedule a dispatch region by using the transform dialect.
-// The schedule is specified by the transform module that is parsed from
-// `transformFileName`.
-std::unique_ptr<Pass> createTransformDialectInterpreterPass(
-    llvm::StringRef transformFileName = llvm::StringRef());
 std::unique_ptr<Pass> createDropSchedulePass();
 } // namespace mlir
